requirements.txt
streamlit
sentence-transformers
google-generativeai
faiss-cpu
numpy
nltk
rouge
